# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Cheaper
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://localhost'
paths:
  /api/user:
    get:
      summary: ''
      operationId: getApiUser
      description: ''
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  id: 25
                  name: Test
                  email: test@gmail.com
                  email_verified_at: null
                  created_at: '2024-09-16T18:17:25.000000Z'
                  updated_at: '2024-09-16T20:22:47.000000Z'
                properties:
                  id:
                    type: integer
                    example: 25
                  name:
                    type: string
                    example: Test
                  email:
                    type: string
                    example: test@gmail.com
                  email_verified_at:
                    type: string
                    example: null
                  created_at:
                    type: string
                    example: '2024-09-16T18:17:25.000000Z'
                  updated_at:
                    type: string
                    example: '2024-09-16T20:22:47.000000Z'
      tags:
        - Endpoints
  /api/products:
    get:
      summary: ''
      operationId: getApiProducts
      description: ''
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                example:
                  -
                    id: 1
                    category_id: 1
                    name: 'Ensemble Gants et Bonnet'
                    slug: ensemble-gants-et-bonnet
                    description: 'couvrez vous'
                    moreDescrciption: 'Restez au chaud à prix tres abordables ,'
                    additionalInfos: '<p>nous pensons aux personnes en situration précaire&nbsp;</p>'
                    stock: 25
                    soldePrice: '3.99'
                    regularPrice: '15.99'
                    imageUrls: '["images\/66e87a590c157_istockphoto-1578054137-612x612.jpg"]'
                    brand: Zora
                    isAvailable: true
                    isBestSeller: true
                    isNewArrival: false
                    isFeatured: false
                    isSpecialOffer: false
                    created_at: '2024-09-16T18:35:05.000000Z'
                    updated_at: '2024-09-16T18:35:05.000000Z'
                  -
                    id: 2
                    category_id: 3
                    name: 'Ensemble Top et jupe'
                    slug: ensemble-top-et-jupe
                    description: 'Magnifique ensemble classe et stylé'
                    moreDescrciption: 'Fashion style'
                    additionalInfos: '<p>top</p>'
                    stock: 15
                    soldePrice: '6.99'
                    regularPrice: '19.99'
                    imageUrls: '["images\/66e87aeae9c81_1.webp"]'
                    brand: HetM
                    isAvailable: false
                    isBestSeller: false
                    isNewArrival: true
                    isFeatured: true
                    isSpecialOffer: false
                    created_at: '2024-09-16T18:37:30.000000Z'
                    updated_at: '2024-09-16T18:37:30.000000Z'
                  -
                    id: 3
                    category_id: 3
                    name: 'Robe de soiree'
                    slug: robe-de-soiree
                    description: 'Robe élégante pour soirée'
                    moreDescrciption: magnifique
                    additionalInfos: '<p>Robe de soiree magnifique&nbsp;</p>'
                    stock: 15
                    soldePrice: '2.95'
                    regularPrice: '17.99'
                    imageUrls: '["images\/66e87bf69c3c1_5.webp"]'
                    brand: EPFC
                    isAvailable: true
                    isBestSeller: false
                    isNewArrival: true
                    isFeatured: true
                    isSpecialOffer: true
                    created_at: '2024-09-16T18:41:58.000000Z'
                    updated_at: '2024-09-16T18:41:58.000000Z'
                  -
                    id: 4
                    category_id: 2
                    name: 'Chemise Blanc et Noir'
                    slug: chemise-blanc-et-noir
                    description: 'Chemise pour Hommes'
                    moreDescrciption: 'soyez élegant à moindre cout'
                    additionalInfos: '<p>……..</p>'
                    stock: 15
                    soldePrice: '4.55'
                    regularPrice: '25.76'
                    imageUrls: '["images\/66e87c6491d80_istockphoto-165685498-612x612.jpg"]'
                    brand: Zarra
                    isAvailable: false
                    isBestSeller: true
                    isNewArrival: true
                    isFeatured: false
                    isSpecialOffer: true
                    created_at: '2024-09-16T18:43:48.000000Z'
                    updated_at: '2024-09-16T18:43:48.000000Z'
                  -
                    id: 5
                    category_id: 2
                    name: 'Ensemble Costume'
                    slug: ensemble-costume
                    description: 'Costme pour homme'
                    moreDescrciption: joli
                    additionalInfos: '<p>…</p>'
                    stock: 3
                    soldePrice: '9.99'
                    regularPrice: '35.99'
                    imageUrls: '["images\/66e87cb8df31d_istockphoto-1430978895-612x612.jpg"]'
                    brand: Primerrrk
                    isAvailable: false
                    isBestSeller: true
                    isNewArrival: true
                    isFeatured: true
                    isSpecialOffer: false
                    created_at: '2024-09-16T18:45:12.000000Z'
                    updated_at: '2024-09-16T18:45:12.000000Z'
                  -
                    id: 6
                    category_id: 3
                    name: 'Jupe simili cuir'
                    slug: jupe-simili-cuir
                    description: 'Affirmez votre style'
                    moreDescrciption: 'Jupe simili cuir'
                    additionalInfos: '<p>…</p>'
                    stock: 15
                    soldePrice: '5.76'
                    regularPrice: '9.99'
                    imageUrls: '["images\/66e87d0d7a1d1_1.webp"]'
                    brand: Zora
                    isAvailable: true
                    isBestSeller: true
                    isNewArrival: true
                    isFeatured: true
                    isSpecialOffer: true
                    created_at: '2024-09-16T18:46:37.000000Z'
                    updated_at: '2024-09-16T18:46:37.000000Z'
                  -
                    id: 7
                    category_id: 3
                    name: Bottines
                    slug: bottines
                    description: "Botine d'hiver"
                    moreDescrciption: Bottes
                    additionalInfos: '<p>…</p>'
                    stock: 15
                    soldePrice: '7.50'
                    regularPrice: '17.70'
                    imageUrls: '["images\/66e87d82665af_istockphoto-1191174327-612x612.jpg"]'
                    brand: Diessel
                    isAvailable: false
                    isBestSeller: true
                    isNewArrival: true
                    isFeatured: false
                    isSpecialOffer: true
                    created_at: '2024-09-16T18:48:34.000000Z'
                    updated_at: '2024-09-16T18:48:34.000000Z'
                  -
                    id: 8
                    category_id: 5
                    name: 'Chaussures de courses Unisex'
                    slug: chaussures-de-courses-unisex
                    description: 'tres léger'
                    moreDescrciption: 'chaussures de courses'
                    additionalInfos: '<p>…</p>'
                    stock: 3
                    soldePrice: '11.50'
                    regularPrice: '24.99'
                    imageUrls: '["images\/66e87e113b266_istockphoto-1453810805-612x612.jpg"]'
                    brand: Abibas
                    isAvailable: true
                    isBestSeller: true
                    isNewArrival: false
                    isFeatured: false
                    isSpecialOffer: true
                    created_at: '2024-09-16T18:50:57.000000Z'
                    updated_at: '2024-09-16T18:50:57.000000Z'
                  -
                    id: 9
                    category_id: 4
                    name: 'Planche à découper'
                    slug: planche-a-decouper
                    description: 'en Bois'
                    moreDescrciption: cuisine
                    additionalInfos: '<p>…</p>'
                    stock: 29
                    soldePrice: '1.50'
                    regularPrice: '5.99'
                    imageUrls: '["images\/66e87e93b729b_istockphoto-2160970399-612x612.jpg"]'
                    brand: Ikoueya
                    isAvailable: true
                    isBestSeller: false
                    isNewArrival: false
                    isFeatured: true
                    isSpecialOffer: true
                    created_at: '2024-09-16T18:53:07.000000Z'
                    updated_at: '2024-09-16T18:53:07.000000Z'
                  -
                    id: 10
                    category_id: 4
                    name: 'Set de Spatules'
                    slug: set-de-spatules
                    description: 'louches et spatules'
                    moreDescrciption: ...
                    additionalInfos: '<p>…</p>'
                    stock: 2
                    soldePrice: '3.29'
                    regularPrice: '6.88'
                    imageUrls: '["images\/66e87edd49f15_istockphoto-1676176264-612x612.jpg"]'
                    brand: Ikoueya
                    isAvailable: true
                    isBestSeller: true
                    isNewArrival: true
                    isFeatured: true
                    isSpecialOffer: true
                    created_at: '2024-09-16T18:54:21.000000Z'
                    updated_at: '2024-09-16T18:54:21.000000Z'
                  -
                    id: 11
                    category_id: 3
                    name: 'Short de sport'
                    slug: short-de-sport
                    description: sportivement
                    moreDescrciption: bien
                    additionalInfos: '<p>tres bien&nbsp;</p>'
                    stock: 3
                    soldePrice: '2.50'
                    regularPrice: '5.99'
                    imageUrls: '["images\/66e87f2249347_6.png"]'
                    brand: Zarra
                    isAvailable: true
                    isBestSeller: true
                    isNewArrival: true
                    isFeatured: true
                    isSpecialOffer: true
                    created_at: '2024-09-16T18:55:30.000000Z'
                    updated_at: '2024-09-16T18:55:30.000000Z'
                  -
                    id: 14
                    category_id: 6
                    name: SmartPhone
                    slug: smartphone
                    description: 'Iphone 11'
                    moreDescrciption: 128g
                    additionalInfos: '<p>Camera frontale 13MP , Camera arriere 20MP</p>'
                    stock: 2
                    soldePrice: '190.55'
                    regularPrice: '399.99'
                    imageUrls: '["images\/6706111ab88c8_iphone11.jpeg"]'
                    brand: Apple
                    isAvailable: true
                    isBestSeller: false
                    isNewArrival: true
                    isFeatured: false
                    isSpecialOffer: false
                    created_at: '2024-10-09T05:14:02.000000Z'
                    updated_at: '2024-10-09T05:49:57.000000Z'
                  -
                    id: 15
                    category_id: 6
                    name: SmartPhone
                    slug: smartphone
                    description: Iphone12
                    moreDescrciption: 64g
                    additionalInfos: '<p>Camera frontale 13MP, Camera arriere 25MP</p>'
                    stock: 3
                    soldePrice: '220.99'
                    regularPrice: '654.77'
                    imageUrls: '["images\/670613c77c394_iphone12.jpeg"]'
                    brand: Apple
                    isAvailable: true
                    isBestSeller: false
                    isNewArrival: true
                    isFeatured: false
                    isSpecialOffer: false
                    created_at: '2024-10-09T05:25:27.000000Z'
                    updated_at: '2024-10-09T05:45:07.000000Z'
                  -
                    id: 16
                    category_id: 6
                    name: 'Pc portable'
                    slug: pc-portable
                    description: 'Dell Latitude occasion'
                    moreDescrciption: 'Pc portable reconditionné 256 giga de mémoire et RAM 8giga'
                    additionalInfos: '<p>redonnez une nouvelles vie a ce pc&nbsp;</p>'
                    stock: 1
                    soldePrice: '99.99'
                    regularPrice: '299.99'
                    imageUrls: '["images\/6706ad617913d_dellLatitude34.jpeg"]'
                    brand: Dell
                    isAvailable: true
                    isBestSeller: false
                    isNewArrival: false
                    isFeatured: false
                    isSpecialOffer: false
                    created_at: '2024-10-09T16:20:49.000000Z'
                    updated_at: '2024-10-09T16:20:49.000000Z'
                  -
                    id: 17
                    category_id: 6
                    name: 'Pc Portable'
                    slug: pc-portable
                    description: 'Asus New generation'
                    moreDescrciption: '256 Gigas de mémoir et 8 Giga de RAM'
                    additionalInfos: '<p>offre de deuxième main&nbsp;</p>'
                    stock: 1
                    soldePrice: '99.99'
                    regularPrice: '320.99'
                    imageUrls: '["images\/6706adf151b1f_asus.jpeg"]'
                    brand: Asus
                    isAvailable: true
                    isBestSeller: false
                    isNewArrival: false
                    isFeatured: false
                    isSpecialOffer: false
                    created_at: '2024-10-09T16:23:13.000000Z'
                    updated_at: '2024-10-09T16:23:13.000000Z'
                  -
                    id: 18
                    category_id: 7
                    name: 'Pc Portable'
                    slug: pc-portable
                    description: 'Thinkpad Lenovo T480'
                    moreDescrciption: '120Giga de memoire , 8 Giga de ram'
                    additionalInfos: '<p>vos articles en seconde main&nbsp;</p>'
                    stock: 2
                    soldePrice: '79.99'
                    regularPrice: '278.96'
                    imageUrls: '["images\/6706bee33fd86_lenovothinkpadT480.jpeg"]'
                    brand: Lenovo
                    isAvailable: true
                    isBestSeller: false
                    isNewArrival: true
                    isFeatured: false
                    isSpecialOffer: false
                    created_at: '2024-10-09T17:35:31.000000Z'
                    updated_at: '2024-10-09T17:35:31.000000Z'
                  -
                    id: 19
                    category_id: 7
                    name: 'Rame de papier'
                    slug: rame-de-papier
                    description: 'format A4'
                    moreDescrciption: 'Rames de papier format A4'
                    additionalInfos: '<p>a moindre cout</p>'
                    stock: 12
                    soldePrice: '2.99'
                    regularPrice: '4.55'
                    imageUrls: '["images\/6706c79b633e8_rame paierA4.jpeg"]'
                    brand: A4
                    isAvailable: true
                    isBestSeller: false
                    isNewArrival: true
                    isFeatured: false
                    isSpecialOffer: false
                    created_at: '2024-10-09T18:12:43.000000Z'
                    updated_at: '2024-10-09T18:12:43.000000Z'
                  -
                    id: 20
                    category_id: 7
                    name: 'Set style , Crayons'
                    slug: set-style-crayons
                    description: 'Stylos  Bic , crayon et Gommes'
                    moreDescrciption: 'indispensables pour les notes'
                    additionalInfos: '<p>notes&nbsp;</p>'
                    stock: 5
                    soldePrice: '0.99'
                    regularPrice: '2.50'
                    imageUrls: '["images\/6706c7f1e3ba4_set stylo.jpeg"]'
                    brand: Bic
                    isAvailable: true
                    isBestSeller: false
                    isNewArrival: true
                    isFeatured: false
                    isSpecialOffer: false
                    created_at: '2024-10-09T18:14:09.000000Z'
                    updated_at: '2024-10-09T18:14:09.000000Z'
                  -
                    id: 21
                    category_id: 7
                    name: 'Sac Pc portable'
                    slug: sac-pc-portable
                    description: 'Sac à dos pour Pc portable'
                    moreDescrciption: 'transpoter votre pc portable en toute sécurité'
                    additionalInfos: '<p>sac à dos&nbsp;</p>'
                    stock: 3
                    soldePrice: '12.99'
                    regularPrice: '25.00'
                    imageUrls: '["images\/6706c8761e825_sac pc portable.jpeg"]'
                    brand: badday
                    isAvailable: true
                    isBestSeller: false
                    isNewArrival: true
                    isFeatured: false
                    isSpecialOffer: false
                    created_at: '2024-10-09T18:16:22.000000Z'
                    updated_at: '2024-10-09T18:16:22.000000Z'
                  -
                    id: 22
                    category_id: 7
                    name: 'Gourde à eau'
                    slug: gourde-a-eau
                    description: 'contenant à eau'
                    moreDescrciption: Contenant
                    additionalInfos: '<p>pour etudiant&nbsp;</p>'
                    stock: 2
                    soldePrice: '9.99'
                    regularPrice: '19.99'
                    imageUrls: '["images\/6706c9d042e71_gourde eau.jpeg"]'
                    brand: Primark
                    isAvailable: true
                    isBestSeller: false
                    isNewArrival: true
                    isFeatured: false
                    isSpecialOffer: false
                    created_at: '2024-10-09T18:19:43.000000Z'
                    updated_at: '2024-10-09T18:22:08.000000Z'
                  -
                    id: 23
                    category_id: 5
                    name: 'Toge Etudiants'
                    slug: toge-etudiants
                    description: 'Pour votre prochaine remise de diplome'
                    moreDescrciption: 'félicitations jeunes diplomés'
                    additionalInfos: '<p>Proficiat , Toge mixte pour remise de diplome&nbsp;</p>'
                    stock: 4
                    soldePrice: '5.99'
                    regularPrice: '25.99'
                    imageUrls: '["images\/6706ce69ad2cf_toge etudiant.jpeg"]'
                    brand: Epfc
                    isAvailable: true
                    isBestSeller: true
                    isNewArrival: true
                    isFeatured: true
                    isSpecialOffer: true
                    created_at: '2024-10-09T18:41:45.000000Z'
                    updated_at: '2024-10-09T18:41:45.000000Z'
      tags:
        - Endpoints
    post:
      summary: ''
      operationId: postApiProducts
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: hhzdf
                description:
                  type: string
                  description: ''
                  example: 'Et delectus ea non quia nulla est ut.'
                price:
                  type: number
                  description: ''
                  example: 0.0
              required:
                - name
                - description
                - price
  '/api/products/{id}':
    get:
      summary: ''
      operationId: getApiProductsId
      description: ''
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  id: 1
                  category_id: 1
                  name: 'Ensemble Gants et Bonnet'
                  slug: ensemble-gants-et-bonnet
                  description: 'couvrez vous'
                  moreDescrciption: 'Restez au chaud à prix tres abordables ,'
                  additionalInfos: '<p>nous pensons aux personnes en situration précaire&nbsp;</p>'
                  stock: 25
                  soldePrice: '3.99'
                  regularPrice: '15.99'
                  imageUrls: '["images\/66e87a590c157_istockphoto-1578054137-612x612.jpg"]'
                  brand: Zora
                  isAvailable: true
                  isBestSeller: true
                  isNewArrival: false
                  isFeatured: false
                  isSpecialOffer: false
                  created_at: '2024-09-16T18:35:05.000000Z'
                  updated_at: '2024-09-16T18:35:05.000000Z'
                properties:
                  id:
                    type: integer
                    example: 1
                  category_id:
                    type: integer
                    example: 1
                  name:
                    type: string
                    example: 'Ensemble Gants et Bonnet'
                  slug:
                    type: string
                    example: ensemble-gants-et-bonnet
                  description:
                    type: string
                    example: 'couvrez vous'
                  moreDescrciption:
                    type: string
                    example: 'Restez au chaud à prix tres abordables ,'
                  additionalInfos:
                    type: string
                    example: '<p>nous pensons aux personnes en situration précaire&nbsp;</p>'
                  stock:
                    type: integer
                    example: 25
                  soldePrice:
                    type: string
                    example: '3.99'
                  regularPrice:
                    type: string
                    example: '15.99'
                  imageUrls:
                    type: string
                    example: '["images\/66e87a590c157_istockphoto-1578054137-612x612.jpg"]'
                  brand:
                    type: string
                    example: Zora
                  isAvailable:
                    type: boolean
                    example: true
                  isBestSeller:
                    type: boolean
                    example: true
                  isNewArrival:
                    type: boolean
                    example: false
                  isFeatured:
                    type: boolean
                    example: false
                  isSpecialOffer:
                    type: boolean
                    example: false
                  created_at:
                    type: string
                    example: '2024-09-16T18:35:05.000000Z'
                  updated_at:
                    type: string
                    example: '2024-09-16T18:35:05.000000Z'
      tags:
        - Endpoints
    put:
      summary: ''
      operationId: putApiProductsId
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: ptdtuxmjhbt
                description:
                  type: string
                  description: ''
                  example: 'Reiciendis aspernatur nobis et adipisci.'
                price:
                  type: number
                  description: ''
                  example: 1035847.0
    delete:
      summary: ''
      operationId: deleteApiProductsId
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
    parameters:
      -
        in: path
        name: id
        description: 'The ID of the product.'
        example: 1
        required: true
        schema:
          type: integer
  /api/generate-token:
    post:
      summary: ''
      operationId: postApiGenerateToken
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
tags:
  -
    name: Endpoints
    description: ''
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: 'You can retrieve your token by visiting your dashboard and clicking <b>Generate API token</b>.'
security:
  -
    default: []